[RENAME] Order Enum 위치 이동 기존 entity 하위에 존재 하던 enum 타입 entity -> enums 패키지 하위로 이동
